requirements.txt
asgiref==3.4.1
astroid==2.6.5
colorama==0.4.4
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
isort==5.9.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.3.1
pylint==2.9.5
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
